Start mockgyver as an application
The purpose behind this is to be able to track concurrent or nestled
mocks better.  There have been problems with the previous approach
that if one testcase crashes and the following starts before mockgyver
has detected that the first crashed ==> mockgyver issues a warning
that nestling mocks is not allowed.

Note that this isn't fully handled yet, but the foundation is now done.