Improve reporting and help in case a waiter waits in vain
Previously calls registered through traces were just dumped when a
WAIT_CALLED timed out.  This has now been improved.  Calls are now
printed in decreasing order of similarity with the pattern specified
in WAIT_CALLED.  Based on an idea by Tomas A and Gustav V.  Thanks.

Not only that, all loaded modules with similar names are checked for
similarly named functions and also printed.

Hopefully this makes it easier for users to spot misspelt module names
or function names or argument mismatches.