Report correct arity for ?WAIT_CALLED patterns
Previously ?WAIT_CALLED resulted in a verification fun (under the
hood) which got all arguments in a list.  This meant that the arity of
the fun was 1 and the help printed to the user couldn't report
expected arity for ?WAIT_CALLED correctly.

    fun([A1, ... An])  ==>  Waiter: .../1

Now, these will be separate arguments and the help is now printed correctly.

    fun(A1, ... An)  ==>  Waiter: .../n