Include filename and line number in error messages
?WAS_CALLED now includes the location (filename and line number) where
the criteria failed (when it does).  This is redundant when using
Erlang/OTP R15B (which provides line numbers in stack traces) or
later, but may prove useful for earlier versions.

?WAIT_CALLED now displays the location which waits for the criteria to
be fulfilled which aids troubleshooting.  The filename and line
numbers are printed in a format which the compilation mode in emacs
renders as clickable links.

Based on ideas from Magnus F and Tomas A.