Steal the trace from the existing tracer (if any)
The problem is that a process may only be traced by one and only one
other process.  We need the traces to record what happens for the
validation afterwards.  One could perhaps design a complicated trace
relay, but at least for the time being we stop the current tracer (if
any) and add ourselves as the sole tracer.