Avoid warning about underscored variables
Erlang/OTP 24 introduces a new warning when an underscored (ignored)
variable is used multiple times in matching statements.
That is, code like this:

    g(_N, _N) -> ok.

Yields a warning like this:

    Warning: variable '_N' is bound multiple times in this pattern. If you mean to ignore this value, use '_' or a different underscore-prefixed name

The mockgyver parse transform would, for a variable N, introduce an
underscored variable ____N (just to make the likelihood of the user
using it in code slim).  The implementation is now changed to name the
variables N@mockgyver which is also unlikely to be used by users.