Load lists of modules at a time
Also cache the code of loaded and mocked modules to be able to
detect when reloading is needed.

Loading lists of modules was measured to be faster on average (see
below for more info).

Instead of iterating over modules on at a time, collect info about
what to load, and load them in a single (mostly) go.  Also save info
for unload as a preparation for a more efficient restoration when the
session ends.  This opens for some later optimizations.

== Measurements for loading list of modules ==

I measured the following for two large modules, let's call
them a and b:

          time
   name   milliseconds      operations measured
  -------------------------------------------------------------
    a      ~120..210        code:purge(a), code:purge(b),
                            code:atomic_load([{a, "", BinA},
                                              {b, "", BinB}])

    b      ~200..210        code:purge(a), code:purge(b),
                            code:atomic_load([a, b])

    c      ~200..210        code:purge(a), code:purge(b),
                            code:load_file(a), code:load_file(b)

    d      ~200..210        code:purge(a), code:purge(b),
                            code:load_binary(a, "", BinA),
                            code:load_binary(b, "", BinB)

The jitter for case a was quite high, but it never took longer than
for the other measurements.  Oddly enough, both cases b and d
were never as fast as case a's fastest.  I repeated each case
a couple of times in the same running Erlang VM.

The beam files were 5.2 and 6.2 MB on disk and carried debug_info.
This was measured with Erlang 23 on an otherwise idling laptop with
an SSD and enough RAM.  The disk I/O ought not to have had impact,
and the files were probably cached.

== Detecting if modules on disk have changed ==

This commit tries to defer code loading to a single step.
To safely do that, and also detect if beam on disk has changed,
it uses beam_lib:md5(File/CodeBinary) to compute the checksum.
This is the same checksum as the Mod:module_info(md5) returns.

Calculating the checksum from file is faster than first loading the
code and then asking for its checksum. Here are my measurements:

   time
   milliseconds  operations
 -------------------------------------------------------------------
  ~120          code:purge(a), code:load_file(a) a:module_info(md5)
  ~35           beam_lib:md5(FilenameOfA)
  ~33           beam_lib:md5(BinaryCodeOfA)

The beam file was 6.2MB on disk and carried debug_info. The file was
the larger one used in the loading measurements above. The sw and hw
setup was the same.

== Other changes ==

Since this tries to defer code loading to as late as possible, it
uses file:read_file/1 and beam_lib:md5/1 for some operations.  I'm
not entirely sure, but this probably means mockgyver with embedded
code loading might not work. I don't know if it did previously or
not though.

This now requires at least Erlang 19 due to the use of
code:ensure_modules_loaded/1 and code:atomic_load/1.

Acked-by: Tomas Abrahamsson <tomas.abrahamsson@gmail.com>
Acked-by: Tomas Abrahamsson <tomas.abrahamsson@gmail.com>