Retain mockings between sessions in a sequence
When setting up tests with a ?WITH_MOCKED_SETUP generator,
don't reload tests between test cases, unless the sequence
appears to be interrupted.

After the sequence, restore the mocks immediately.

This can speed eunit ?WITH_MOCKED_SETUP generators up considerably.

It works like this:

* In ?WITH_MOCKED_SETUP, set the mock_sequence option
* When a session in a sequence ends, don't restore mocked modules
* When the last session in a sequence ends, do restore mocked modules
* When a session in a sequence starts, don't re-setup mocked modules
* Use a test signature to detect if a new mock sequence starts
* If the next session in a sequence doesn't started quickly enough,
  restore mocked modules, since something might be wrong.

A backwards incompatible change is that it no longer works to mock the
'application' module with ?WITH_MOCKED_SETUP, since the exec/4 depends
on it.  The retaining of mockings between sessions collides with
possibilities of mocking the 'application' module.