Redirect to mod^ if mock is called between sessions
If a function in a mocked module is called _between_ sessions,
redirect the call to the renamed module^ so that the "original" code
gets called.

This could happen if the user/programmer mocks a module that is
needed by eunit or by mockgyver between sessions, or a module
that is central to Erlang.