Introduce elements of a session
A session element is a part of a session, and a session can consist
of one or several session elements.

Extend the state machine with one more state, the 'session_element'
state in addition to the already existing 'no_session' and 'session'
states to model this.

This allows for a performance optimization by loading and unloading
the mocks only when the session starts and ends.